Move stopwatch implementation to StopWatch.tsx
